+++ foodiehotspots/tasks.py
@@ -7,9 +7,7 @@ from django_apscheduler.jobstores import DjangoJobStore, register_events, regist
 
 from django.db import models
 # from foodiehotspots.views import RestaurantScheduler
-from foodiehotspots.schedulers import RestaurantScheduler
-from foodiehotspots.scheduler import DiscordWebHooksScheduler
-import time
+from foodiehotspots.scheduler import RestaurantScheduler, DiscordWebHooksScheduler
 
 
 scheduler1 = None
@@ -23,11 +21,9 @@ def start():
     # 이전 스케줄이 완료될 때까지 대기할 시간(초) 설정
     wait_time = 1000
     job1 = partial(RestaurantScheduler.restaurant_scheduler, RestaurantScheduler())
-    # job2 = partial(RestaurantCacheScheduler.set_foodhotspots, RestaurantCacheScheduler())
-    # scheduler.add_job(job2, 'cron', hour='3', misfire_grace_time=wait_time)  # 2시 실행(default)
-    # scheduler.add_job(job2, 'cron', second='*/20', misfire_grace_time=wait_time)  # 2시 실행(default)
-    # scheduler.add_job(job1, 'cron', minute='*',misfire_grace_time=wait_time) #!매분 실행으로 테스트
+    
     scheduler1.add_job(job1, 'cron', hour=2, misfire_grace_time=wait_time)  # 2시 실행(default)
+    
     # scheduler1.add_job(job1, 'cron', minute='*',misfire_grace_time=wait_time) #!매분 실행으로 테스트
     scheduler1.start()
